refactor: combine preservation.rs and persistence.rs
