+++ tests/end_to_end_cases/mod.rs
@@ -5,7 +5,6 @@ pub mod management_cli;
 pub mod operations_api;
 pub mod operations_cli;
 mod persistence;
-pub mod preservation;
 pub mod read_api;
 pub mod read_cli;
 pub mod scenario;
